Update CardFetcher.java
I would update this to fetch card info from Yugioh Fandom to Yugipedia. Yugipedia is updated more often and has better images. On yugioh fandom, some of the passcodes aren't setup properly so if you use this it misses a lot of cards out. Some examples are eldlich and drytron. Some new decks.